refactor : 페이지 이름을 index에서 page로 일관성 있게 통일
