--- src/pages/register/components/index.tsx
@@ -1,5 +1,0 @@
-const RegisterPage = () => {
-  return <div>회원가입</div>;
-};
-
-export default RegisterPage;
