+++ src/pages/subscriptions/components/SubscriptionPage.tsx
@@ -1,0 +1,3 @@
+export const SubscriptionPage = () => {
+  return <div></div>;
+};
